+++ src/js/Node.js
@@ -98,7 +98,7 @@ export class Node {
         const getValue = this.getValue.bind(this)
         const editable = this.editor.options.onEditable({
           field: this.field,
-          get value() {
+          get value () {
             return getValue()
           },
           path: this.getPath()
@@ -951,7 +951,7 @@ export class Node {
       const addClasses = this.editor.options.onClassName({
         path: this.getPath(),
         field: this.field,
-        get value() {
+        get value () {
           return getValue()
         }
       }) || ''
@@ -1562,7 +1562,7 @@ export class Node {
    */
   _setValueError (message) {
     this.valueError = {
-      message: message
+      message
     }
     this.updateError()
   }
@@ -1581,7 +1581,7 @@ export class Node {
    */
   _setFieldError (message) {
     this.fieldError = {
-      message: message
+      message
     }
     this.updateError()
   }
@@ -1617,8 +1617,8 @@ export class Node {
       path: this.getInternalPath(),
       oldValue: this.previousValue,
       newValue: this.value,
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
 
     this.previousValue = this.value
@@ -1650,8 +1650,8 @@ export class Node {
       index: this.getIndex(),
       oldValue: this.previousField,
       newValue: this.field,
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
 
     this.previousField = this.field
@@ -1741,12 +1741,7 @@ export class Node {
           this.dom.select.appendChild(defaultOption)
 
           // Iterate all enum values and add them as options
-          for (let i = 0; i < this.enum.length; i++) {
-            const option = document.createElement('option')
-            option.value = this.enum[i]
-            option.textContent = this.enum[i]
-            this.dom.select.appendChild(option)
-          }
+          this._updateEnumOptions()
 
           this.dom.tdSelect = document.createElement('td')
           this.dom.tdSelect.className = 'jsoneditor-tree'
@@ -1773,7 +1768,7 @@ export class Node {
           delete this.valueFieldHTML
         }
       } else {
-        // cleanup select box when displayed
+        // cleanup select box when displayed, and attach the editable div instead
         if (this.dom.tdSelect) {
           this.dom.tdSelect.parentNode.removeChild(this.dom.tdSelect)
           delete this.dom.tdSelect
@@ -1781,6 +1776,8 @@ export class Node {
           this.dom.tdValue.innerHTML = this.valueFieldHTML
           this.dom.tdValue.style.visibility = ''
           delete this.valueFieldHTML
+
+          this.dom.tdValue.appendChild(this.dom.value)
         }
       }
 
@@ -1849,6 +1846,23 @@ export class Node {
       stripFormatting(domValue)
 
       this._updateDomDefault()
+    }
+  }
+
+  _updateEnumOptions () {
+    if (!this.enum || !this.dom.select) {
+      return
+    }
+
+    // clear the existing options
+    this.dom.select.innerHTML = ''
+
+    // Iterate all enum values and add them as options
+    for (let i = 0; i < this.enum.length; i++) {
+      const option = document.createElement('option')
+      option.value = this.enum[i]
+      option.textContent = this.enum[i]
+      this.dom.select.appendChild(option)
     }
   }
 
@@ -2260,6 +2274,7 @@ export class Node {
         domField.innerHTML = escapedField
       }
       this._updateSchema()
+      this._updateEnumOptions()
     }
 
     // apply value to DOM
@@ -2872,12 +2887,12 @@ export class Node {
               oldIndex: oldNextNode.getIndex(),
               newIndex: firstNode.getIndex(),
 
-              oldIndexRedo: oldIndexRedo,
-              newIndexRedo: newIndexRedo,
-              oldParentPathRedo: oldParentPathRedo,
-              newParentPathRedo: newParentPathRedo,
+              oldIndexRedo,
+              newIndexRedo,
+              oldParentPathRedo,
+              newParentPathRedo,
 
-              oldSelection: oldSelection,
+              oldSelection,
               newSelection: this.editor.getDomSelection()
             })
           }
@@ -2931,12 +2946,12 @@ export class Node {
             oldIndex: oldNextNode.getIndex(),
             newIndex: firstNode.getIndex(),
 
-            oldIndexRedo: oldIndexRedo,
-            newIndexRedo: newIndexRedo,
-            oldParentPathRedo: oldParentPathRedo,
-            newParentPathRedo: newParentPathRedo,
+            oldIndexRedo,
+            newIndexRedo,
+            oldParentPathRedo,
+            newParentPathRedo,
 
-            oldSelection: oldSelection,
+            oldSelection,
             newSelection: this.editor.getDomSelection()
           })
         }
@@ -2978,12 +2993,12 @@ export class Node {
               oldIndex: oldNextNode.getIndex(),
               newIndex: firstNode.getIndex(),
 
-              oldIndexRedo: oldIndexRedo,
-              newIndexRedo: newIndexRedo,
-              oldParentPathRedo: oldParentPathRedo,
-              newParentPathRedo: newParentPathRedo,
+              oldIndexRedo,
+              newIndexRedo,
+              oldParentPathRedo,
+              newParentPathRedo,
 
-              oldSelection: oldSelection,
+              oldSelection,
               newSelection: this.editor.getDomSelection()
             })
           }
@@ -3048,13 +3063,13 @@ export class Node {
             fieldNames: selectedNodes.map(getField),
             oldParentPath: oldParent.getInternalPath(),
             newParentPath: firstNode.parent.getInternalPath(),
-            oldParentPathRedo: oldParentPathRedo,
-            newParentPathRedo: newParentPathRedo,
-            oldIndexRedo: oldIndexRedo,
-            newIndexRedo: newIndexRedo,
+            oldParentPathRedo,
+            newParentPathRedo,
+            oldIndexRedo,
+            newIndexRedo,
             oldIndex: oldNextNode.getIndex(),
             newIndex: firstNode.getIndex(),
-            oldSelection: oldSelection,
+            oldSelection,
             newSelection: this.editor.getDomSelection()
           })
         }
@@ -3096,6 +3111,14 @@ export class Node {
       // Put the table online again
       frame.appendChild(table)
       frame.scrollTop = scrollTop
+    }
+
+    if (typeof this.editor.options.onExpand === 'function') {
+      this.editor.options.onExpand({
+        path: this.getPath(),
+        isExpand: this.expanded,
+        recursive: recurse
+      })
     }
   }
 
@@ -3154,7 +3177,7 @@ export class Node {
     const newNode = new Node(this.editor, {
       field: (field !== undefined) ? field : '',
       value: (value !== undefined) ? value : '',
-      type: type
+      type
     })
     newNode.expand(true)
 
@@ -3168,10 +3191,10 @@ export class Node {
     this.editor._onAction('insertBeforeNodes', {
       nodes: [newNode],
       paths: [newNode.getInternalPath()],
-      beforePath: beforePath,
+      beforePath,
       parentPath: this.parent.getInternalPath(),
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
   }
 
@@ -3188,7 +3211,7 @@ export class Node {
     const newNode = new Node(this.editor, {
       field: (field !== undefined) ? field : '',
       value: (value !== undefined) ? value : '',
-      type: type
+      type
     })
     newNode.expand(true)
     this.parent.insertAfter(newNode, this)
@@ -3201,8 +3224,8 @@ export class Node {
       paths: [newNode.getInternalPath()],
       afterPath: this.getInternalPath(),
       parentPath: this.parent.getInternalPath(),
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
   }
 
@@ -3219,7 +3242,7 @@ export class Node {
     const newNode = new Node(this.editor, {
       field: (field !== undefined) ? field : '',
       value: (value !== undefined) ? value : '',
-      type: type
+      type
     })
     newNode.expand(true)
     this.parent.appendChild(newNode)
@@ -3231,8 +3254,8 @@ export class Node {
       nodes: [newNode],
       paths: [newNode.getInternalPath()],
       parentPath: this.parent.getInternalPath(),
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
   }
 
@@ -3250,10 +3273,10 @@ export class Node {
 
       this.editor._onAction('changeType', {
         path: this.getInternalPath(),
-        oldType: oldType,
-        newType: newType,
-        oldSelection: oldSelection,
-        newSelection: newSelection
+        oldType,
+        newType,
+        oldSelection,
+        newSelection
       })
     }
   }
@@ -3320,7 +3343,7 @@ export class Node {
     if (triggerAction === true) {
       this.editor._onAction('sort', {
         path: this.getInternalPath(),
-        oldChilds: oldChilds,
+        oldChilds,
         newChilds: this.childs
       })
     }
@@ -3337,7 +3360,7 @@ export class Node {
 
     this.editor._onAction('transform', {
       path: this.getInternalPath(),
-      oldValue: oldValue,
+      oldValue,
       newValue: this.getInternalValue()
     })
   }
@@ -3362,8 +3385,8 @@ export class Node {
     this.hide({ resetVisibleChilds: false })
 
     return {
-      table: table,
-      nextTr: nextTr
+      table,
+      nextTr
     }
   }
 
@@ -3954,7 +3977,7 @@ export class Node {
 
       items = this.editor.options.onCreateMenu(items, {
         type: 'single',
-        path: path,
+        path,
         paths: [path]
       })
     }
@@ -4115,7 +4138,7 @@ export class Node {
             path: this.getPath(),
             size: count,
             type: this.type,
-            get value() {
+            get value () {
               return getValue()
             }
           })
@@ -4214,7 +4237,7 @@ Node.onDragStart = (nodes, event) => {
     oldParentPathRedo: parent.getInternalPath(),
     oldIndexRedo: firstNode.getIndex(),
     mouseX: event.pageX,
-    offsetY: offsetY,
+    offsetY,
     level: firstNode.getLevel()
   }
   document.body.style.cursor = 'move'
@@ -4429,14 +4452,14 @@ Node.onDragEnd = (nodes, event) => {
       count: nodes.length,
       fieldNames: nodes.map(getField),
 
-      oldParentPath: oldParentPath,
-      newParentPath: newParentPath,
-      oldIndex: oldIndex,
-      newIndex: newIndex,
+      oldParentPath,
+      newParentPath,
+      oldIndex,
+      newIndex,
 
-      oldIndexRedo: oldIndexRedo,
-      newIndexRedo: newIndexRedo,
-      oldParentPathRedo: oldParentPathRedo,
+      oldIndexRedo,
+      newIndexRedo,
+      oldParentPathRedo,
       newParentPathRedo: null, // This is a hack, value will be filled in during undo
 
       oldSelection: editor.drag.oldSelection,
@@ -4527,7 +4550,7 @@ Node._findSchema = (topLevelSchema, schemaRefs, path, currentSchema = topLevelSc
           if (segment in currentSchema) {
             currentSchema = currentSchema[segment]
           } else {
-            throw Error(`Unable to resovle reference ${ref}`)
+            throw Error(`Unable to resolve reference ${ref}`)
           }
         }
       } else if (ref.match(/#\//g)?.length === 1) {
@@ -4614,12 +4637,12 @@ Node.onRemove = nodes => {
 
     // store history action
     editor._onAction('removeNodes', {
-      nodes: nodes,
-      paths: paths,
+      nodes,
+      paths,
       parentPath: parent.getInternalPath(),
       index: firstIndex,
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
   }
 }
@@ -4675,8 +4698,8 @@ Node.onDuplicate = nodes => {
       clonePaths: clones.map(getInternalPath),
       afterPath: lastNode.getInternalPath(),
       parentPath: parent.getInternalPath(),
-      oldSelection: oldSelection,
-      newSelection: newSelection
+      oldSelection,
+      newSelection
     })
   }
 }