+++ src/js/previewmode.js
@@ -1,6 +1,6 @@
 'use strict'
 
-import jsonrepair from 'jsonrepair'
+import { jsonrepair } from 'jsonrepair'
 import {
   DEFAULT_MODAL_ANCHOR,
   MAX_PREVIEW_CHARACTERS,
@@ -256,8 +256,12 @@ previewmode.create = function (container, options = {}) {
     }
   }
 
+  const errorTableVisible = Array.isArray(this.options.showErrorTable)
+    ? this.options.showErrorTable.includes(this.mode)
+    : this.options.showErrorTable === true
+
   this.errorTable = new ErrorTable({
-    errorTableVisible: true,
+    errorTableVisible,
     onToggleVisibility: function () {
       me.validate()
     },